requirements.txt
awsebcli==3.7.5
blessed==1.9.5
botocore==1.4.10
cement==2.8.2
click==6.6
colorama==0.3.7
django-cors-headers==1.1.0
django==1.9.5
djangorestframework==3.3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
pathspec==0.3.4
pip==7.1.2
pls==0.0.5
python-dateutil==2.5.2
pyyaml==3.11
requests==2.9.1
semantic-version==2.5.0
setuptools==20.6.7
six==1.10.0
texttable==0.8.4
virtualenv==15.0.1
wcwidth==0.1.6
websocket-client==0.35.0
wheel==0.24.0
